+++ test/PhpParser/PrettyPrinterTest.php
@@ -261,10 +261,6 @@ CODE
          * This test makes sure that the format-preserving pretty printer round-trips for all
          * the pretty printer tests (i.e. returns the input if no changes occurred).
          */
-        if (false !== strpos($code, 'new class')) {
-            // Can't preserve formatting on anon classes for now
-            return;
-        }
 
         list($version) = $this->parseModeLine($modeLine);